refactor: Improve marks module consistency and simplify mark identification
- Add comprehensive documentation for RadiusExpression and radius_expression methods
- Extract is_continuous_scale helper using scale.range_kind() for extensibility
- Move utility functions (is_continuous_scale, default_scale_for_data_type) to marks/util.rs
- Simplify mark identification from pointer-based strings to index-based usize
- Remove mark_id() method from Mark trait, use enumeration indices instead
- Fix clippy warnings (unnecessary casts, collapsible matches, derivable impls)
- Clean up unused imports in test files

This refactor improves code organization, makes scale checking extensible for external crates,
and simplifies the legend merging system by using deterministic mark indices.

+++ avenger-chart/src/legend_renderer/symbol.rs
@@ -351,9 +351,9 @@ impl LegendRenderer for SymbolLegendRenderer {
             ScalarOrArray::new_scalar(crate::utils::parse_color_string_strict(&default_fill)?);
         legend_config.stroke =
             ScalarOrArray::new_scalar(crate::utils::parse_color_string_strict(&default_stroke)?);
-        legend_config.size = ScalarOrArray::new_scalar(default_size as f32);
-        legend_config.angle = ScalarOrArray::new_scalar(default_angle as f32);
-        legend_config.stroke_width = Some(default_stroke_width as f32);
+        legend_config.size = ScalarOrArray::new_scalar(default_size);
+        legend_config.angle = ScalarOrArray::new_scalar(default_angle);
+        legend_config.stroke_width = Some(default_stroke_width);
 
         // Process each channel
         for channel in channels {
